Ported to IDA7 (Qt, etc.)

+++ sync-plugin/include/sync/SyncManager.h
@@ -12,7 +12,7 @@ class SyncManager
 	ISyncHandler* m_syncHandler[NumSyncHandlers];
 
 	// Notifications
-	static int __stdcall ida_notification_point(void*, int notificationCode, va_list args);
+	static ssize_t __stdcall ida_notification_point(void*, int notificationCode, va_list args);
 	bool m_notificationLock = false;
 
 	void OnIdaNotification(IdaNotification&);
